Move QyCodeDimensionControl to qr-codes sub-folder
